refactor: folder structure
