+++ app/src/store/slices/tabs-layout/index.ts
@@ -1,0 +1,2 @@
+export * from "./slice";
+export * from "./selectors";
