+++ app/src/store/slices/tabs-layout/selectors.ts
